Fix pidUpdate() being called 3x more often than necessary in controllerCorrectRatePID().
TRUNCATE_SINT16() was a preprocessor macro that evaluates its input three times in the common case.  Where it was used, the input is a call to pidUpdate() with updateError true, so in addition to the performance impact, there are unintended side effects.  Specifically, the integral would wind up 3x more then intended, and the derivative constant would have no effect at all, other than to potentially break the TRUNCATE_SINT16() logic.  Luckily the three PID loops affected by this did not use the derivative constant.

I've replaced the macro with an inline function, removed an unused version of the macro, and divided the integral and integral windup terms of the three affected PID loops by 3 to compensate.  The resulting behavior should be unchanged.