Enable fno-math-errno
C standard library math routines are defined as setting errno in certain
cases.  For applications like this firmware, this is bad, for two
reasons:

1. It means that certain hardware floating point operations,
   particularly VSQRT, must be wrapped with functions that check flags
   and set errno.  This can easily reduce their performance by 5-10x.

2. It introduces greater reliance on errno, which (particularly in
   FreeRTOS) is difficult to link in a thread-safe manner.  The default
   configuration, and indeed the one used by this firmware, has all
   threads writing to a *shared* errno variable!

However, this firmware isn't *checking* errno, so we don't have to put
up with all this.  This commit tells the compiler that we're not
expecting errno side effects from math functions, freeing it to directly
use hardware instructions like VSQRT.

As a result, we can resolve the TODOs around errno linkage and remove
the declaration of the shared errno variable.