radiolink: create queue before starting syslink.
If the STM32 reboots while a connection is active, the nRF will usually
have data waiting on the UART as soon as syslink is ready to process it.
Because the priority of the syslink task is higher than the task calling
radiolinkInit (namely, the system task), syslink will immediately
preempt.

Net effect: it tries to stuff a packet into the crtpPacketDelivery queue
and FreeRTOS asserts.

Here I've ensured that the shared queue is created before the client
task starts up, and added an assert to catch the (unlikely) case where
allocating the queue fails.  This makes finding the culprit
significantly easier, since the alternative is to trigger an assert
somewhere deep within FreeRTOS.

STM32 reboots with an RF connection active are now robust.

Note that this is a mitigation measure; there may be more bugs like
this.