maxsonar: fix apparent unit errors.
The MB1040 accuracy function was taking a measurement in millimeters,
multiplying it by 25.4, and comparing it with inches.  This appears to
have been a simple inversion of the desired math, which I've fixed.

I've also switched it to single-precision arithmetic; this escaped my
notice in my earlier batch of fixes because it's not compiled by
default.