Closes #72: Fixed freeze when connecting USB
When connecting with USB there is a 1 second pause caused by the
client (see bitcraze/crazyflie-clients/python#213). During this
pause no packets are transfered and this resulted in an interrup
storm of the 'tx fifo empty' interrupt. It freezes the Copter and
became visible when we enabled the watchdog (see #71).

This commit is not a real fix but more a workaround. It makes sure
there is always at least one packet to transfer to the PC for every
USB frame. This is done in the SOF interrupt, a NULL CRTP packet is
sent if nothing else has to be sent. This way the interrupt storm
is avoided.