Enables all EXTI handlers to use callbacks
This commit centrally enables all EXTI interrupts on the NVIC. Note that the individual EXTI line interrupts still need to be enabled for interrupts to be generated.

This fixes issues related to EXTI9_5 and EXTI15_10 handlers being initialized and defined multiple times.

This affects, primarily, the IMU interrupt (EXTI13) and interrupts on some of the deck pins (eg. dw1000 on EXTI11). The solution was therefore made global, and existing interrupts updated to use callbacks, rather than the handlers directly.

Signed-off-by: Mike Hamer <mike@mikehamer.info>