Sensors read by interrupt-driven task
This commit introduces an interrupt-driven task to handle IMU reads using DMA.

The task configures the IMU's master mode to handle barometer and magnetometer access.

The task waits on an IMU Data Ready interrupt on EXTI13 before initializing a DMA read (via CPAL) of the IMU's registers.

Since the IMU handles barometer and magnetometer access, all sensors can be read in a single DMA transaction. This eliminates I2C contention problems, since only one transaction occurs on the bus (rather than separately reading from 3 sensors on one bus).

This significantly improves performance of the sensors. sensors_stock.c takes 700us to read the gyro and accelerometer data (most of the time spent waiting), and thus severely limits the rate of the stabilizer task. This was never an issue since the stock stabilizer is computationally light, but becomes important once more advanced onboard estimators or controllers are added to the mix.

IMPORTANT NOTE: As of this commit, the magnetometer only updates once, and its values stay constant in the IMU registers. I don't know why, but I also don't know anyone who uses a magnetometer for indoor flight... so this is a low priority for me to fix. I have disabled the magnetometer for the time being.

Signed-off-by: Mike Hamer <mike@mikehamer.info>