Update motors_def_cf2.c
First of all the comment in line 402 claims to use Timer 3 Channel 1 for MOSI on PA7. However PA7 only has acces to Timer 3 Channel 2! 
Either the comment needs to be changed, or an other timer should be used. Since CPPM is unlikely wanted to be used when someone changes the motormapping to PA7, I propose to use timer 14 channel 1 in the motordefinition.

Furthermore the default motor mapping of the Bigquad Deck as described in lines 437 to 440 was wrong and should be replaced with my changes.