Updated unit tests (rake) to be called from make.
The reason is that make manages the build configuration (defines and so on) and we want to reuse them in the unit tests. When rake is called from make, all defines (actually compiler flags) are passed on to rake to be used when building the unit tests.
Also added a light weight “annotation” that allows unit test files to be excluded based on defines.