#211 Fixed vl53 deck driver hijacking i2c buss.
The vl53 deck driver used interrupt polling and a periodic timer to read
out sensor data. The periodic timer was running faster then the vl53
sensor could provide values leaving the driver in a constant polling
state. This had the consequence that the mem subsystem, having a lower
priority, never got to read the eeprom on the same i2c buss. A
relaxation delay was added to fix this. Also the sensor reading was
changed from every 100ms to as fast as possible.