Eanbled crash/tubled detection. Closes #248
I the sitaw module the crash/tumbled detection was enabled in case the
kalman estimator is used. Currently it is automatically used for the
flow and the LPS decks. It checks if the attitude angle is above 60 deg
for a period of time. If so shuts of motors until it comes back below 60
deg again.