Move syslink packet parsing into ISR
Handle the parsing of syslink packets in ISR and post complete packets
to a queue. Client syslink blocks on this queue and is only woken when a
full packet is available.

This eliminates a context switch on every character received and
possibly allows for lowering of the syslink task priority without
risking dropped/corrupt packets.