HighLevelCmd:Broadcast support for startTrajectory
The previous design of startTrajectory required to pass in information
about the memory layout. When using multiple quadrotors with different
trajectories, this design does not allow to start trajectories at the same
time using a broadcast.

The updated API design now has a defineTrajectory method where the
memory layout can be defined for a trajectoryId. StartTrajectory
just requires the trajectoryId. CFs can now be configured individually
with their trajectories, and trajectory execution can be started
simulatanously using broadcasts.

Verified with 2 CFs and a modified Crazyswarm package using a figure8
trajectory.