LocSrv: Add EXT_POS_PACKED CRTP packet
The packed version enables to use very efficient broadcasts in a mocap system
to support a large quantity of CFs. This CRTP packet can carry position
information for up to 4 CFs (in a space of -16...16m due to compression).

Tested with a modified Crazyswarm package and 2 CFs in a VICON Mocap space.

Note that this can support twice as many CFs per packet as the original
Crazyswarm packet, because no orientations are transmitted. This
should allow us to use about 30 CFs per radio (not tested).